Move JcTools to sdk-common
