+++ integration-tests/graal/build.gradle.kts
@@ -9,14 +9,14 @@ otelJava.moduleName.set("io.opentelemetry.graal.integration.tests")
 sourceSets {
   main {
     // We need to ensure that we have the shadowed classes on the classpath, without this
-    // we will get the <:sdk:trace-shaded-deps> classes only, without the shadowed ones
-    val traceShadedDeps = project(":sdk:trace-shaded-deps")
-    output.dir(traceShadedDeps.file("build/extracted/shadow"), "builtBy" to ":sdk:trace-shaded-deps:extractShadowJar")
+    // we will get the <:sdk:common-shaded-deps> classes only, without the shadowed ones
+    val traceShadedDeps = project(":sdk:common-shaded-deps")
+    output.dir(traceShadedDeps.file("build/extracted/shadow"), "builtBy" to ":sdk:common-shaded-deps:extractShadowJar")
   }
 }
 
 dependencies {
-  implementation(project(path = ":sdk:trace-shaded-deps"))
+  implementation(project(path = ":sdk:common-shaded-deps"))
 }
 
 graalvmNative {
