+++ sdk-extensions/autoconfigure/build.gradle.kts
@@ -25,7 +25,7 @@ dependencies {
 
   annotationProcessor("com.google.auto.value:auto-value")
 
-  testImplementation(project(":sdk:trace-shaded-deps"))
+  testImplementation(project(":sdk:common-shaded-deps"))
 
   testImplementation(project(":sdk:testing"))
   testImplementation("com.linecorp.armeria:armeria-junit5")
@@ -85,7 +85,7 @@ testing {
         implementation(project(":exporters:prometheus"))
         implementation(project(":exporters:zipkin"))
         implementation(project(":sdk:testing"))
-        implementation(project(":sdk:trace-shaded-deps"))
+        implementation(project(":sdk:common-shaded-deps"))
         implementation(project(":semconv"))
 
         implementation("com.google.guava:guava")
