+++ sdk-extensions/incubator/build.gradle.kts
@@ -14,7 +14,7 @@ otelJava.moduleName.set("io.opentelemetry.sdk.extension.incubator")
 dependencies {
   api(project(":sdk:all"))
 
-  compileOnly(project(":sdk:trace-shaded-deps"))
+  compileOnly(project(":sdk:common-shaded-deps"))
 
   annotationProcessor("com.google.auto.value:auto-value")
