+++ sdk/common-shaded-deps/src/main/resources/META-INF/native-image/io.opentelemetry/opentelemetry-sdk-common/reflect-config.json
