+++ sdk/common-shaded-deps/src/test/java/io/opentelemetry/sdk/common/internal/JcToolsSecurityManagerTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.trace.internal;
+package io.opentelemetry.sdk.common.internal;
 
 import static org.assertj.core.api.Assertions.assertThat;
