+++ settings.gradle.kts
@@ -54,13 +54,13 @@ include(":opentracing-shim")
 include(":perf-harness")
 include(":sdk:all")
 include(":sdk:common")
+include(":sdk:common-shaded-deps")
 include(":sdk:logs")
 include(":sdk:logs-testing")
 include(":sdk:metrics")
 include(":sdk:metrics-testing")
 include(":sdk:testing")
 include(":sdk:trace")
-include(":sdk:trace-shaded-deps")
 include(":sdk-extensions:autoconfigure")
 include(":sdk-extensions:autoconfigure-spi")
 include(":sdk-extensions:incubator")
